server/myplay/requirements.txt
Django==2.1.1
dj-database-url==0.5.0
uwsgi==2.0.17.1
python-dotenv==0.7.1
psycopg2-binary==2.7.4
